split userIcon from NavBar.tsx
